#58: Rename LaplacianSmoother file
