+++ movement/__init__.py
@@ -1,5 +1,5 @@
 from movement.tangling import *  # noqa
 
-from movement.laplacian import *  # noqa
+from movement.laplacian_smoothing import *  # noqa
 from movement.monge_ampere import *  # noqa
 from movement.spring import *  # noqa
